+++ RobWork/src/rwlibs/algorithms/RANSACModel.hpp
@@ -1,7 +1,7 @@
 /********************************************************************************
- * Copyright 2009 The Robotics Group, The Maersk Mc-Kinney Moller Institute, 
- * Faculty of Engineering, University of Southern Denmark 
- * 
+ * Copyright 2009 The Robotics Group, The Maersk Mc-Kinney Moller Institute,
+ * Faculty of Engineering, University of Southern Denmark
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -34,7 +34,7 @@ namespace algorithms {
 
 /**
  * @brief An interface for RANSAC model fitting.
- * 
+ *
  * @todo A model needs to remember the indices of inliers from the set of data...
  */
 template<class MODEL, class DATA>
@@ -82,7 +82,7 @@ public:
 
 		int n = MODEL().getMinReqData();
 
-		if (data.size() < n || data.size() < dataRequired) {
+		if (data.size() < static_cast<size_t>(n) || data.size() < static_cast<size_t>(dataRequired)) {
 			//RW_WARN("Too few samples to create a proper model.");
 
 			return std::vector<MODEL>();
@@ -99,7 +99,7 @@ public:
 		while (iterations++ < maxIterations) {
 			std::random_shuffle(indices.begin(), indices.end());
 			std::vector<DATA> maybeInliers;
-			for (size_t i = 0; i < n; ++i) {
+			for (size_t i = 0; i < static_cast<size_t>(n); ++i) {
 				maybeInliers.push_back(data[indices[i]]);
 			}
 
@@ -126,7 +126,7 @@ public:
 			}
 
 			// if consensus set size is large enough, we have a model
-			if (consensusSet.size() > dataRequired) {
+			if (consensusSet.size() > static_cast<size_t>(dataRequired)) {
 				models.push_back(
 						std::pair<MODEL, int>(maybeModel, consensusSet.size()));
 			}
@@ -156,7 +156,7 @@ public:
 
 				if (models[0].first.invalid()
 						|| models[0].first.getNumberOfInliers()
-								< dataRequired) {
+								< static_cast<size_t>(dataRequired)) {
 					return std::vector<MODEL>();
 				}
 			} catch (...) {
@@ -177,7 +177,7 @@ public:
 		bool merging = false;
 		do {
 			merging = false;
-			
+
 			for (size_t i = 0; i < modelsPtr.size() - 1; i++) {
 
 				if (modelsPtr[i].first == NULL) {
@@ -207,7 +207,7 @@ public:
 
 					// mark the model as processed
 					modelsPtr[j].first = NULL;
-					
+
 					merging = true;
 				}
 
@@ -219,7 +219,7 @@ public:
 				std::vector<DATA> consensusSet;
 				std::vector<size_t> consensusSetIndices;
 				for (size_t k = 0; k < data.size(); ++k) {
-					
+
 					if (bestCloseModel.first->belongsTo(data[k], dataThreshold)) {
 						consensusSet.push_back(data[k]);
 						consensusSetIndices.push_back(k);
@@ -231,7 +231,7 @@ public:
 					bestCloseModel.first->_indices = consensusSetIndices;
 
 					if (bestCloseModel.first->invalid()
-						|| bestCloseModel.first->getNumberOfInliers() < dataRequired
+						|| bestCloseModel.first->getNumberOfInliers() < static_cast<size_t>(dataRequired)
 					) {
 						continue;
 					}
@@ -239,10 +239,10 @@ public:
 					continue;
 				}
 
-				
+
 			}
 		} while (merging);
-		
+
 		std::vector<MODEL> newModels;
 		for (size_t i = 0; i < modelsPtr.size() - 1; i++) {
 			if (modelsPtr[i].first != NULL) {
@@ -440,7 +440,7 @@ protected:
 	std::vector<DATA> _data;
 	double _quality;
 
-	/* 
+	/*
 	 * For some calculations, it is neccesary to know which data points given to findModels()
 	 * function were actualy fitted in the model. This vector holds the indices of inliers.
 	 * Use at your own risk, if you know what you are doing.
